# doc/api.yml
openapi: 3.0.3
info:
  title: Texas Hold'em Poker API
  description: |
    A secure API for Texas Hold'em poker, enforcing a strict split between Table Management (Games) and Deal Execution.
  version: 1.0.0
servers:
  - url: /poker
    description: Base path for the poker game

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Game:
      type: object
      properties:
        gameId:
          type: string
          format: uuid
        name:
          type: string
        blinds:
          $ref: '#/components/schemas/Blinds'
        players:
          type: array
          items:
            $ref: '#/components/schemas/PlayerInfo'
          description: |
            Excludes anyone who has left the table (see POST /games/{gameId}/leave), except a
            player who left while all-in in a hand still in progress - they stay visible until
            that hand reaches showdown.
        dealHistory:
          type: array
          items:
            type: string
            format: uuid
          description: List of historical deal IDs played at this table.
        currentDealId:
          type: string
          format: uuid
          nullable: true
        currentDeal:
          $ref: '#/components/schemas/DealState'
        you:
          $ref: '#/components/schemas/YouState'
        closed:
          type: boolean
          description: Once true, the table is permanently done - no more joins, deals, or actions.
        defaultBuyIn:
          type: integer
          format: int64
          nullable: true
          description: |
            The buy-in amount the table's creator used when they sat down, kept as a suggested
            default for anyone else buying in later. Null if the creator didn't set one.
      required:
        - gameId
        - blinds
        - players
        - dealHistory
        - closed

    YouState:
      type: object
      nullable: true
      description: |
        Betting context personalized for the authenticated caller. Only present when the request
        included a valid Authorization header for a player seated in this game, and a deal is
        currently in progress. Absent (or the whole "you" key omitted) for unauthenticated/preview
        requests, spectators, or when there is no active deal.
      properties:
        callAmount:
          type: integer
          format: int64
          description: Additional chips needed to call and match the current round's bet.
        minRaise:
          type: integer
          format: int64
          description: Minimum legal amount for a BET or RAISE right now.
        maxRaise:
          type: integer
          format: int64
          description: The caller's current chip stack (an all-in raise/bet).
      required:
        - callAmount
        - minRaise
        - maxRaise

    PlayerInfo:
      type: object
      properties:
        playerId:
          type: string
          format: uuid
        username:
          type: string
        chips:
          type: integer
          format: int64
        status:
          $ref: '#/components/schemas/PlayerStatus'
        contributionThisRound:
          type: integer
          format: int64
          description: Chips this player has put in during the current betting round.
        blind:
          $ref: '#/components/schemas/BlindRole'
        lastAction:
          type: string
          nullable: true
          description: |
            Human-readable description of the last action this player took in the current betting
            round (e.g. "CALL 20", "FOLD", "CHECK"). Null if they haven't acted since the last
            street change.
      required:
        - playerId
        - username
        - chips
        - status
        - contributionThisRound

    PlayerStatus:
      type: string
      enum: [ACTIVE, FOLDED, ALL_IN]

    BlindRole:
      type: string
      enum: [SMALL, BIG]
      nullable: true
      description: Which blind this player posted this deal, or null if neither.

    Blinds:
      type: object
      properties:
        small:
          type: integer
          format: int64
        big:
          type: integer
          format: int64
      required:
        - small
        - big

    DealState:
      type: object
      properties:
        dealId:
          type: string
          format: uuid
        communityCards:
          type: array
          items:
            $ref: '#/components/schemas/CommunityCard'
          minItems: 5
          maxItems: 5
          description: |
            A 5-card array representing the table deck.
            Initiates empty (all nulls) and populates progressively as rounds transition.
        pot:
          type: integer
          format: int64
        phase:
          $ref: '#/components/schemas/Phase'
        activePlayerId:
          type: string
          format: uuid
          nullable: true
        currentRoundBet:
          type: integer
          format: int64
          description: The highest contribution any player has made in the current betting round.
        outcome:
          $ref: '#/components/schemas/DealOutcome'
        turnDeadline:
          type: string
          format: date-time
          nullable: true
          description: |
            Server-computed deadline for activePlayerId to act, derived from the last action taken
            on this deal plus a fixed window (DealService.TURN_TIMEOUT). Null once there is no
            active player (e.g. past showdown). If a player's turn reaches this deadline without
            them acting, the server auto-folds them the next time this game is read or acted on.
      required:
        - dealId
        - communityCards
        - pot
        - phase
        - currentRoundBet

    DealOutcome:
      type: object
      nullable: true
      description: Present once the deal has reached SHOWDOWN and the pot has been awarded.
      properties:
        winners:
          type: array
          items:
            $ref: '#/components/schemas/WinnerInfo'
        revealedHands:
          type: array
          description: |
            Every non-folded player's hole cards and hand rank at a real (card-based) showdown,
            winners and losers alike. Empty when the pot was won uncontested (every other player
            folded), since hands are never compared or revealed in that case.
          items:
            $ref: '#/components/schemas/RevealedHand'
      required:
        - winners
        - revealedHands

    WinnerInfo:
      type: object
      properties:
        playerId:
          type: string
          format: uuid
        username:
          type: string
        amount:
          type: integer
          format: int64
          description: Chips this winner was awarded from the pot.
        handRank:
          type: string
          nullable: true
          description: |
            Human-readable hand description, e.g. "two pair, Js and 9s". Null when the pot was won
            uncontested (every other player folded), since hands are never compared or revealed
            in that case.
      required:
        - playerId
        - username
        - amount

    RevealedHand:
      type: object
      properties:
        playerId:
          type: string
          format: uuid
        username:
          type: string
        holeCards:
          type: array
          items:
            type: string
          description: This player's two hole cards, e.g. ["AS", "TD"].
        handRank:
          type: string
          description: Human-readable hand description, e.g. "two pair, Js and 9s".
      required:
        - playerId
        - username
        - holeCards
        - handRank

    Card:
      type: string
      pattern: '^([2-9]|10|[JQKA])[cdhsCDHS]$'
      example: 'As'
      description: |
        Compact card representation.
        Rank: 2-10, J, Q, K, A.
        Suit: c (Clubs), d (Diamonds), h (Hearts), s (Spades).

    CommunityCard:
      allOf:
        - $ref: '#/components/schemas/Card'
      nullable: true
      description: A community card which can be null if not yet revealed.

    Phase:
      type: string
      enum: [PRE_FLOP, FLOP, TURN, RIVER, SHOWDOWN]

    TimelineEvent:
      type: object
      properties:
        type:
          type: string
          enum: [PLAYER_JOINED, DEAL_STARTED, PHASE_CHANGED, ACTION_TAKEN, SHOWDOWN, POT_AWARDED]
        timestamp:
          type: string
          format: date-time
        payload:
          type: object
      required:
        - type
        - timestamp

    ActionRequest:
      type: object
      properties:
        actionType:
          type: string
          enum: [CHECK, CALL, FOLD, RAISE, BET]
        amount:
          type: integer
          format: int64
          description: Required for RAISE and BET.
      required:
        - actionType

    Hand:
      type: object
      properties:
        cards:
          type: array
          items:
            $ref: '#/components/schemas/Card'
          minItems: 2
          maxItems: 2
      required:
        - cards

    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message

    AuthRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 3
        password:
          type: string
          minLength: 6
      required:
        - username
        - password

    TokenResponse:
      type: object
      properties:
        token:
          type: string
          description: A signed JWT to send as `Authorization Bearer <token>` on subsequent requests.
      required:
        - token

security:
  - BearerAuth: []

paths:
  /auth/register:
    post:
      summary: Create an account and receive a signed auth token
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Username already taken
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/login:
    post:
      summary: Exchange username/password for a signed auth token
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Invalid username or password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /games:
    post:
      summary: Create a new game table
      tags: [Game Infrastructure]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                blinds:
                  $ref: '#/components/schemas/Blinds'
                defaultBuyIn:
                  type: integer
                  format: int64
                  nullable: true
                  description: |
                    Optional. Kept as the table's suggested default buy-in for anyone joining
                    later (see Game.defaultBuyIn).
              required:
                - name
                - blinds
      responses:
        '201':
          description: Game created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /games/{gameId}:
    get:
      summary: Get game details
      description: |
        Authorization is optional here: omit it for an unauthenticated preview (e.g. before
        joining), or include a valid Bearer token to also receive the personalized "you" betting
        context if you're seated in this game. An invalid (malformed/unsigned) token still fails
        with 401 - "optional" only covers omitting the header entirely.
      tags: [Game Infrastructure]
      security:
        - {}
        - BearerAuth: []
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Game details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '404':
          description: Game not found
        '401':
          $ref: '#/components/responses/Unauthorized'

  /games/{gameId}/players:
    post:
      summary: Register/Buy-in a player to the game
      tags: [Game Infrastructure]
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                buyInAmount:
                  type: integer
                  format: int64
              required:
                - buyInAmount
      responses:
        '204':
          description: Player registered successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /games/{gameId}/close:
    post:
      summary: Close the table for good
      description: |
        Ends the table permanently: no further joins, deals, or actions are accepted afterwards.
        The game remains readable (final chip counts stay visible). Any player currently seated
        at the table may close it - there is no separate host/owner role.
      tags: [Game Infrastructure]
      security:
        - BearerAuth: []
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Table closed
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Caller is not seated at this table
        '409':
          $ref: '#/components/responses/Conflict'

  /games/{gameId}/leave:
    post:
      summary: Stand up from the table for good
      description: |
        The caller is excluded from every future deal at this table (they no longer count toward
        the players list or the minimum-2-players check) and, if a hand is currently in progress,
        folded out of it unless they're already folded, all-in (an all-in player has no more
        decisions to make by leaving, so they stay in for that hand's showdown), or the hand has
        already reached showdown. Not reversible - there is no way to sit back down.

        A player who has been auto-folded on turn timeout for several hands in a row without ever
        acting is treated the same way automatically (see DealState.turnDeadline) - this endpoint
        covers the manual case.
      tags: [Game Infrastructure]
      security:
        - BearerAuth: []
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Left the table
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Caller is not seated at this table
        '409':
          $ref: '#/components/responses/Conflict'

  /games/{gameId}/deals:
    post:
      summary: Start a new deal at the table
      description: |
        Shuffles the deck, automatically posts the small and big blind, and deals two hole
        cards to each player with chips. Requires at least 2 players with chips to start.
      tags: [Game Infrastructure]
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '201':
          description: Deal started; returns the updated game state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'

  /games/{gameId}/events:
    get:
      summary: Stream of timeline events for the game
      tags: [Game Infrastructure]
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: SSE stream of TimelineEvents
          content:
            text/event-stream:
              schema:
                $ref: '#/components/schemas/TimelineEvent'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /deals/{dealId}/actions:
    post:
      summary: Take a gameplay action
      tags: [Deal Execution]
      parameters:
        - name: dealId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActionRequest'
      responses:
        '204':
          description: Action accepted
        '400':
          description: Illegal move or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Acting out of turn or unauthorized access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          $ref: '#/components/responses/Conflict'

  /deals/{dealId}/hand/me:
    get:
      summary: Fetch the calling player's private cards
      tags: [Deal Execution]
      parameters:
        - name: dealId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: The player's hand
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Hand'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Unauthorized hand access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

responses:
  BadRequest:
    description: Bad Request
    content:
      application/json:
        schema:
          $ref: '#/components/schemas/Error'
  Unauthorized:
    description: Unauthorized
    content:
      application/json:
        schema:
          $ref: '#/components/schemas/Error'
  Conflict:
    description: Conflict
    content:
      application/json:
        schema:
          $ref: '#/components/schemas/Error'